% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediccion.R
\name{predictML}
\alias{predictML}
\title{Predicciones de un modelo de regresión.}
\usage{
predictML(x_nuevo, modelo_entrenado)
}
\arguments{
\item{x_nuevo}{(vector) datos de la variable explicativa para los cuales se requiere la predicción.}

\item{modelo_entrenado}{salida de la función \code{regresion}.}
}
\value{
Un vector de valores predichos.
}
\description{
Realiza las predicciones de un modelo de regresión lineal simple.
}
\examples{
\dontrun{
# Cargando la librería
library(RegLinS)

# Entreno mi modelo
x <- c(1,2,3,4,5)
y <- c(5, 5.4, 6, 7, 9)
salida <- regresion(y = y, x = x)

# Realizo predicciones
xnew <- c(2.7, 4.5, 3.1)
predictML(x_nuevo = xnew, modelo_entrenado = salida)
}
}
