% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_lineal.R
\name{regresion}
\alias{regresion}
\title{Regresión lineal simple}
\usage{
regresion(y, x)
}
\arguments{
\item{y}{(vector) datos de la respuesta.}

\item{x}{(vector) datos de la variable explicativa.}
}
\value{
Una lista con los coeficientes de regresión, errores estándar, intervalo de confianza, y coeficiente de determinación.
}
\description{
Realiza una regresión lineal simple.
}
\examples{
\dontrun{
# Directorio de trabajo
ruta <- "~/ruta/test.csv"

# -----------------------------------
# Ejemplo 1:
MisDatos <- read.csv(ruta)
# Cargo la librería
library(RegLinS)
regresion(y = MisDatos$peso, x = MisDatos$ancho_corazon)

# -----------------------------------
# Ejemplo 2:
x <- c(1,2,3,4,5)
y <- c(5, 5.4, 6, 7, 9)

regresion(y = y, x = x)
}
}
